backend/requirements.txt
fastapi
uvicorn
pandas
langchain
langchain-groq
pydantic
python-dotenv
